requirements.txt
asgiref==3.3.4
Django==3.2.3
django-environ==0.4.5
djangorestframework==3.11.0
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
